sql,settings: move TestLintClusterSettingNames to the settings pkg
Release note: None
