+++ pkg/BUILD.bazel
@@ -312,6 +312,7 @@ ALL_TESTS = [
     "//pkg/server/tenantsettingswatcher:tenantsettingswatcher_test",
     "//pkg/server/tracedumper:tracedumper_test",
     "//pkg/server:server_test",
+    "//pkg/settings/lint:lint_test",
     "//pkg/settings/rulebasedscanner:rulebasedscanner_test",
     "//pkg/settings:settings_test",
     "//pkg/spanconfig/spanconfigbounds:spanconfigbounds_test",
@@ -1586,6 +1587,7 @@ GO_TARGETS = [
     "//pkg/server:server",
     "//pkg/server:server_test",
     "//pkg/settings/cluster:cluster",
+    "//pkg/settings/lint:lint_test",
     "//pkg/settings/rulebasedscanner:rulebasedscanner",
     "//pkg/settings/rulebasedscanner:rulebasedscanner_test",
     "//pkg/settings:settings",
